mk1/huggingface/requirements.txt
gradio==5.24.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
transformers>=4.11.0
Pillow>=8.3.1
numpy>=1.21.0 